Fix: clone assumed to be in home (#84)
* Fix bug clone assumed to be in home

The scriplet clones in cwd but assumes cwd is ~ in the next command.

Minor: removed the rm -rf of the repo, only purpose it seems to have is
to make debugging harder

* Don't ignore errors in the scriplet

Errors are ignored so even if the checkout fails, this appears to pass
because the installation is carried out, so one has to squint to notice.
Lets fail on error